Universal node: nodesonosPlayer, node removed in signitares
