+++ src/Discovery.js
@@ -12,7 +12,7 @@
 const { PACKAGE_PREFIX } = require('./Globals.js')
 
 const { getGroupsAll: xGetGroupsAll } = require('./Sonos-CommandsTs.js')
-const { matchSerialUuid: xMatchSerialUuid } = require('./HelperTs.js')
+const { matchSerialUuid: xMatchSerialUuid } = require('./Sonos-Extensions.js')
 
 const { SonosDeviceDiscovery, SonosDevice } = require('@svrooij/sonos/lib')
