+++ src/Sonos-Extensions.js
@@ -17,8 +17,9 @@ const { PACKAGE_PREFIX } = require('./Globals.js')
 const { executeActionV6 } = require('./Sonos-Commands.js')
 
 const { isTruthyPropertyStringNotEmpty: xIsTruthyPropertyStringNotEmpty,
+  isTruthyStringNotEmpty: xIsTruthyStringNotEmpty,
   isTruthyArray: xIsTruthyArray, isTruthy: xIsTruthy, isTruthyProperty: xIsTruthyProperty,
-  encodeHtmlEntity: xEncodeHtmlEntity
+  encodeHtmlEntity: xEncodeHtmlEntity, decodeHtmlEntity: xDecodeHtmlEntity
 } = require('./HelperTs.js')
 
 const  request   = require('axios').default
@@ -33,7 +34,7 @@ module.exports = {
   //
 
   /** Validate that ip belongs to a SONOS player.
-   * @param {object} playerUrlObject player URL JavaScript build in Object 
+   * @param {object} playerUrlObject player JavaScript build-in URL 
    * @param {number} timeout in milliseconds
    * 
    * @returns {Promise<boolean>} true if typical SONOS player response
@@ -64,7 +65,7 @@ module.exports = {
   },
 
   /** Get device properties.
-   * @param {object} playerUrlObject player URL JavaScript build in Object 
+   * @param {object} playerUrlObject player JavaScript build-in URL
    *
    * @returns {Promise<object>} device properties as object
    *
@@ -156,7 +157,7 @@ module.exports = {
    * CAUTION: Allows to set all to empty - means no content, will not play
    * Thats done for the createSnapshot, restoreSnapshot!
    * 
-   * @param {object} playerUrlObject player URL build in JavaScript object
+   * @param {object} playerUrlObject player JavaScript build-in URL 
    * @param {object} inArgs action arguments (except InstanceID)
    * @param {string} inArgs.CurrentURI such as "x-sonosapi-stream:s119032?sid=254&flags=8224&sn=0"
    * @param {string} inArgs.CurrentURIMetaData uri as DIDL-Lite xml
@@ -204,7 +205,7 @@ module.exports = {
   //
   //                    SIMPLE COMMANDS - EXECUTE ACTION AND SIMPLE TRANSFORMATION
   //
-  //  @param  {object} playerUrlObject/coordinatorUrlObject player URL JavaScript build in Object
+  //  @param  {object} playerUrlObject/coordinatorUrlObject player JavaScript build-in URL urlObject
   //  @returns always a promise
   //  @throws {error} from executeAction
   //
@@ -311,4 +312,179 @@ module.exports = {
       { 'InstanceID': 0, 'Channel': 'Master', 'DesiredVolume': newVolume })
   },
 
+  /** Comparing player UUID and serial number. Returns true if matching.
+   * @param  {string} serial the string such as 00-0E-58-FE-3A-EA:5
+   * @param  {string} uuid the string such as RINCON_000E58FE3AEA01400
+   * RINCONG_xxxxxxxxxxxx01400  (01400 is port)
+   * 
+   * @returns {Promise<boolean>} true if matching
+   * 
+   * @throws only split, replace exceptions
+   * 
+   * Algorithm: only checks the first part of serial number
+   * 
+   * @since 2021-02-13
+   */
+  matchSerialUuid: (serial, uuid) => {
+    debug('method >>%s', 'matchSerialUuid')
+    
+    let serialClean = serial.split(':')[0]
+    serialClean = serialClean.replace(/-/g, '')
+
+    let uuidClean = uuid.replace(/^(RINCON_)/, '')
+    uuidClean = uuidClean.replace(/(01400)$/, '')
+    
+    return (uuidClean === serialClean)
+  },
+
+  /** 
+   * Returns an array of items (DidlBrowseItem) extracted from action "Browse" output.
+   * @param  {string} browseDidlXml DIDL-Light string in xml format from Browse (original!)
+   * @param  {string}  itemName DIDL-Light property holding the data. Such as "item" or "container"
+   *
+   * @returns {Promise<DidlBrowseItem[]>} Promise, array of {@link Sonos-CommandsTs#DidlBrowseItem},
+   *  maybe empty array.
+   *                   
+   * @throws {error} if any parameter is missing
+   * @throws {error} from method xml2js and invalid response (missing id, title)
+   * 
+   * Browse provides the results (property Result) in form of a DIDL-Lite xml format. 
+   * The <DIDL-Lite> includes several attributes such as xmlns:dc" and entries 
+   * all named "container" or "item". These include xml tags such as 'res'. 
+   */
+  parseBrowseDidlXmlToArray: async function (browseDidlXml, itemName) {
+    if (!xIsTruthyStringNotEmpty(browseDidlXml)) {
+      throw new Error(`${PACKAGE_PREFIX} DIDL-Light input is missing`)
+    }
+    if (!xIsTruthyStringNotEmpty(itemName)) {
+      throw new Error(`${PACKAGE_PREFIX} item name such as container is missing`)
+    }
+    const decoded = await xDecodeHtmlEntity(browseDidlXml)
+    const didlJson = await parser.parse(decoded, {
+      'ignoreAttributes': false,
+      'attributeNamePrefix': '_',
+      'parseNodeValue': false // this is important - example Title 49 will otherwise be converted
+    })  
+    if (!xIsTruthyProperty(didlJson, ['DIDL-Lite'])) {
+      throw new Error(`${PACKAGE_PREFIX} invalid response Browse: missing DIDL-Lite`)
+    }
+
+    let originalItems = []
+    // single items are not of type array (fast-xml-parser)
+    const path = ['DIDL-Lite', itemName]
+    if (xIsTruthyProperty(didlJson, path)) {
+      const itemsOrOne = didlJson[path[0]][path[1]]
+      if (Array.isArray(itemsOrOne)) { 
+        originalItems = itemsOrOne.slice()
+      } else { // single item  - convert to array
+        originalItems.push(itemsOrOne) 
+      }
+    } 
+
+    // transform properties Album related
+    const transformedItems = await Promise.all(originalItems.map(async (item) => {
+      const newItem = {
+        'id': '',
+        'title': '',
+        'artist': '',
+        'album': '',
+        'uri': '',
+        'artUri': '',
+        'metadata': '',
+        'sid': '',
+        'serviceName': '',
+        'upnpClass': '',
+        'processingType': 'queue' // has to be updated in calling program
+      }
+      if (!xIsTruthyProperty(item, ['_id'])) {
+        throw new Error(`${PACKAGE_PREFIX} id is missing`) // should never happen
+      }
+      newItem.id = item['_id']
+
+      if (!xIsTruthyProperty(item, ['dc:title'])) {
+        throw new Error(`${PACKAGE_PREFIX} title is missing`) // should never happen
+      }
+      if (xIsTruthyProperty(item, ['dc:creator'])) {
+        newItem.artist = item['dc:creator']
+      }
+      if (!xIsTruthyProperty(item, ['dc:title'])) {
+        throw new Error(`${PACKAGE_PREFIX} title is missing`) // should never happen
+      }
+      newItem.title = await xDecodeHtmlEntity(String(item['dc:title'])) // clean title for search
+      if (xIsTruthyProperty(item, ['dc:creator'])) {
+        newItem.artist = item['dc:creator']
+      }
+      if (xIsTruthyProperty(item, ['res', '#text'])) {
+        newItem.uri = item['res']['#text'] // HTML entity encoded, URI encoded
+        newItem.sid = module.exports.getMusicServiceId(newItem.uri)
+        newItem.serviceName = module.exports.getMusicServiceName(newItem.sid)
+      }
+      if (xIsTruthyProperty(item, ['upnp:class'])) {
+        newItem.upnpClass = item['upnp:class']
+      }
+      // artURI (cover) maybe an array (one for each track) then choose first
+      let artUri = ''
+      if (xIsTruthyProperty(item, ['upnp:albumArtURI'])) {
+        artUri = item['upnp:albumArtURI']
+        if (Array.isArray(artUri)) {
+          if (artUri.length > 0) {
+            newItem.artUri = artUri[0]
+          }
+        } else {
+          newItem.artUri = artUri
+        }
+      }
+      // special case My Sonos favorites. It include metadata in DIDL-lite format.
+      // these metadata include the original title, original upnp:class (processingType)
+      if (xIsTruthyProperty(item, ['r:resMD'])) {
+        newItem.metadata = item['r:resMD']
+      }
+      return newItem
+    })
+    )
+    return transformedItems  // properties see transformedItems definition
+  },
+
+  /**  Get music service id (sid) from Transport URI.
+   * @param  {string} xuri such as (masked)
+   * "x-rincon-cpcontainer:1004206ccatalog%2falbums%***%2f%23album_desc?sid=201&flags=8300&sn=14"
+   *
+   * @returns {string} service id or if not found empty string
+   *
+   * prerequisites: uri is string where the sid is in between "?sid=" and "&flags="
+   */
+  getMusicServiceId: (xuri) => {
+    debug('method >>%s', 'getMusicServiceId')
+    let sid = '' // default even if uri undefined.
+    if (xIsTruthyStringNotEmpty(xuri)) {
+      const positionStart = xuri.indexOf('?sid=') + '$sid='.length
+      const positionEnd = xuri.indexOf('&flags=')
+      if (positionStart > 1 && positionEnd > positionStart) {
+        sid = xuri.substring(positionStart, positionEnd)
+      }
+    }
+    return sid
+  },
+
+  /**  Get service name for given service id.
+   * @param  {string} sid service id (integer) such as "201" or blank 
+   * 
+   * @returns {string} service name such as "Amazon Music" or empty string
+   *
+   * @uses database of services (map music service id  to musics service name)
+   */
+  getMusicServiceName: (sid) => {
+    debug('method >>%s', 'getMusicServiceName')
+    let serviceName = '' // default even if sid is blank
+    if (sid !== '') {
+      const list = module.exports.MUSIC_SERVICES
+      const index = list.findIndex((service) => {
+        return (service.sid === sid)
+      })
+      if (index >= 0) {
+        serviceName = list[index].name
+      }  
+    } 
+    return serviceName
+  }
 }
