+++ src/sonos-universal.js
@@ -290,7 +290,7 @@ module.exports = function (RED) {
     }
     
     // eslint-disable-next-line max-len
-    return COMMAND_TABLE_UNIVERSAL[command](node, msg, nodesonosPlayer, tsPlayer)
+    return COMMAND_TABLE_UNIVERSAL[command](msg, tsPlayer)
   }
 
   //
@@ -299,18 +299,18 @@ module.exports = function (RED) {
   
   /**
    *  Coordinator delegate coordination of group. New player must be in same group!
-   * @param {object} node not used
+
    * @param {object} msg incoming message
    * @param {string} msg.payload new coordinator name - must be in same group and different
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
   // eslint-disable-next-line max-len
-  async function coordinatorDelegateCoordination (node, msg, nodesonosPlayer, tsPlayer) {
+  async function coordinatorDelegateCoordination (msg, tsPlayer) {
     // Payload new player name is required.
     const validPlayerName = validRegex(msg, 'payload', REGEX_ANYCHAR, 'player name', PACKAGE_PREFIX)
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
@@ -342,17 +342,16 @@ module.exports = function (RED) {
 
   /**
    *  Adjust group volume and outputs new volume.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {(string|number)} msg.payload -100 to + 100, integer
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {Promise<String>} Returns the new group volume after adjustment as property newVolume.
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupAdjustVolume (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupAdjustVolume (msg, tsPlayer) {
     // Payload adjusted volume is required
     const adjustVolume = validToInteger(msg, 'payload', -100, +100, 'adjust volume', PACKAGE_PREFIX)
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
@@ -366,16 +365,15 @@ module.exports = function (RED) {
 
   /**
    *  Clear queue.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupClearQueue (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupClearQueue (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     // eslint-disable-next-line max-len
@@ -388,18 +386,17 @@ module.exports = function (RED) {
 
   /**
    *  Create a snapshot of the given group of players.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {boolean} [msg.snapVolumes = false] will capture the players volumes
    * @param {boolean} [msg.snapMutestates = false] will capture the players mutestates
    * @param {string} [msg.playerName=using tssPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {payload: snap} snap see xCreateGroupSnapshot
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupCreateSnapshot (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupCreateSnapshot (msg, tsPlayer) {
     debug('entering method groupCreateSnapshot')
     // Validate msg properties
     const options = { 'snapVolumes': false, 'snapMutestates': false } // Default
@@ -426,17 +423,16 @@ module.exports = function (RED) {
 
   /**
    *  Group create volume snap shot (used for adjust group volume)
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
   // eslint-disable-next-line max-len
-  async function groupCreateVolumeSnapshot (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupCreateVolumeSnapshot (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     await executeActionV6(groupData.members[0].urlObject, // coordinator at 0
@@ -448,16 +444,15 @@ module.exports = function (RED) {
 
   /**
    *  Cancel group sleep timer.
-   * @param {object} node not used
    * @param {object} msg incoming message.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupCancelSleeptimer (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupCancelSleeptimer (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     await executeActionV6(groupData.members[0].urlObject, // coordinator at 0
@@ -469,17 +464,16 @@ module.exports = function (RED) {
 
   /**
    *  Get group transport actions.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {payload: transportActions}
    *
    * @throws any functions throws error and explicit throws
    */
   // eslint-disable-next-line max-len
-  async function groupGetTransportActions (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupGetTransportActions (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const payload = await executeActionV6(groupData.members[0].urlObject, // coordinator at 0
@@ -491,16 +485,15 @@ module.exports = function (RED) {
 
   /**
    *  Get group crossfade mode.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {payload: crossfade mode} on|off
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupGetCrossfadeMode (node, msg, sonosPlayer, tsPlayer) {
+  async function groupGetCrossfadeMode (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const state = await executeActionV6(groupData.members[0].urlObject, // coordinator at 0
@@ -512,16 +505,15 @@ module.exports = function (RED) {
 
   /**
    *  Get array of group member - this group.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {Promise<GroupMember[]>} with key payload!
    *
    * @throws {error} from methods validatedGroupProperties, getGroupsCurrent
    */
-  async function groupGetMembers (node, msg, sonosPlayer, tsPlayer) {
+  async function groupGetMembers (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
 
@@ -530,16 +522,15 @@ module.exports = function (RED) {
 
   /**
    *  Get group mute.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise<string>} on|off
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupGetMute (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupGetMute (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const state = await executeActionV6(groupData.members[0].urlObject, // coordinator at 0
@@ -551,10 +542,9 @@ module.exports = function (RED) {
 
   /**
    *  Get the playback state of that group, the specified player belongs to.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise<string>} state
    * state: { 'stopped', 'playing', 'paused_playback', 'transitioning', 'no_media_present' }
@@ -563,28 +553,27 @@ module.exports = function (RED) {
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupGetPlaybackstate (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupGetPlaybackstate (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
-    const tsCoordinator = new Sonos(groupData.members[0].urlObject.hostname) 
+    const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname) 
     tsCoordinator.urlObject = groupData.members[0].urlObject
-    const payload = xGetPlaybackstate(tsCoordinator.urlObject)
+    const payload = await xGetPlaybackstate(tsCoordinator.urlObject)
 
     return { payload }
   }
 
   /**
    *  Get group SONOS queue - the SONOS queue of the coordinator.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise<object>} object to update msg. msg.payload = array of queue items as object
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupGetQueue (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupGetQueue (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname) 
@@ -596,16 +585,15 @@ module.exports = function (RED) {
 
   /**
    *  Get group sleeptimer.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {payload: crossfade mode} hh:mm:ss
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupGetSleeptimer (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupGetSleeptimer (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const sleep = await executeActionV6(groupData.members[0].urlObject, // coordinator at 0
@@ -623,10 +611,9 @@ module.exports = function (RED) {
 
   /**
    *  Get state (see return) of that group, the specified player belongs to.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject- as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject- as default
    *
    * @returns {promise} { see return }
    * state: 'stopped' | 'playing' | 'paused_playback' | 'transitioning' | 'no_media_present' }
@@ -635,7 +622,7 @@ module.exports = function (RED) {
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupGetState (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupGetState (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname)
@@ -691,17 +678,16 @@ module.exports = function (RED) {
 
   /**
    *  Get group track, media and position info.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {payload: media: {object}, trackInfo: {object}, 
    *  positionInfo: {object}, queueActivated: true/false
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupGetTrackPlus (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupGetTrackPlus (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname)
@@ -795,16 +781,15 @@ module.exports = function (RED) {
 
   /**
    *  Get group volume.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise<string>} volume
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupGetVolume (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupGetVolume (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const payload = await executeActionV6(groupData.members[0].urlObject, // coordinator at 0
@@ -816,16 +801,15 @@ module.exports = function (RED) {
 
   /**
    *  Play next track on given group of players.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupNextTrack (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupNextTrack (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname)
@@ -836,16 +820,15 @@ module.exports = function (RED) {
 
   /**
    *  Pause playing in that group, the specified player belongs to.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupPause (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPause (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname)
@@ -856,18 +839,17 @@ module.exports = function (RED) {
 
   /**
    *  Starts playing content. Content must have been set before.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
    * @param {number} [msg.sameVolume=true] shall all players play at same volume level. 
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupPlay (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPlay (msg, tsPlayer) {
     // Validate msg.playerName, msg.volume, msg.sameVolume -error are thrown
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
@@ -893,7 +875,6 @@ module.exports = function (RED) {
 
   /**
    *  Play data being exported form My Sonos (uri/metadata) on a given group of players
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload content to be played
    * @param {string} msg.payload.uri uri to be played/queued
@@ -902,14 +883,14 @@ module.exports = function (RED) {
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
    * @param {boolean} [msg.sameVolume=true] shall all players play at same volume level.
    * @param {boolean} [msg.clearQueue=true] if true and export.queue = true the queue is cleared.
-   * @param {string} [msg.playerName=using nodesonosPlayer] SONOS-Playername
-   * @param {object} nodesonosPlayer player with urlObject - as default
+   * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
+   * @param {object} tsPlayer node-sonos player with urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws  any functions throws error and explicit throws
    */
-  async function groupPlayExport (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPlayExport (msg, tsPlayer) {
     // Simple validation of export and activation
 
     const exportData = msg.payload
@@ -966,7 +947,6 @@ module.exports = function (RED) {
 
   /**
    *  Play notification on a given group of players. Group topology will not being touched.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload notification uri.
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
@@ -974,7 +954,7 @@ module.exports = function (RED) {
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
    * @param {string} [msg.duration] duration of notification hh:mm:ss 
    *  - default is calculation, if that fails then 00:00:05
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
@@ -984,7 +964,7 @@ module.exports = function (RED) {
    *  While playing a notification (start .. to end + 2 seconds)
    *     there should not be send another request to this group.
    */
-  async function groupPlayNotification (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPlayNotification (msg, tsPlayer) {
     // Payload uri is required.
     const validatedUri = validRegex(msg, 'payload', REGEX_ANYCHAR, 'uri', PACKAGE_PREFIX)
 
@@ -1026,18 +1006,17 @@ module.exports = function (RED) {
 
   /**
    *  Play none empty queue.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
    * @param {number} [msg.sameVolume=true] shall all players play at same volume level. 
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupPlayQueue (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPlayQueue (msg, tsPlayer) {
     // Validate msg.playerName, msg.volume, msg.sameVolume -error are thrown
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
@@ -1072,17 +1051,16 @@ module.exports = function (RED) {
 
   /**
    *  Play a given snapshot on the given group of players.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload snapshot - output form groupCreateSnapshot
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupPlaySnapshot (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPlaySnapshot (msg, tsPlayer) {
     if (xIsTruthyProperty(msg, ['payload'])) {
       if (typeof msg.payload !== 'object') {
         throw new Error(`${PACKAGE_PREFIX}: snapshot (msg.payload) is not object`)
@@ -1119,13 +1097,12 @@ module.exports = function (RED) {
 
   /**
    *  Plays stream using http such as http://www.fritz.de/live.m3u, https://live.radioarabella.de
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload uri start with http(s):// 
    * @param {(number|string)} [msg.volume=unchanged] new volume
    * @param {boolean} [msg.sameVolume=true] force all players to play at same volume level.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {Promise<boolean>} always true
    * 
@@ -1133,7 +1110,7 @@ module.exports = function (RED) {
    * @throws {error} NRCSP error validatedGroupProperties, getGroupCurrent, validRegex
    * @throws {error} from node-sonos setAVTransportURI and setPlayerVolume
    */
-  async function groupPlayStreamHttpV2 (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPlayStreamHttpV2 (msg, tsPlayer) {
     // Payload uri is required.
     let validatedUri = validRegex(msg, 'payload', REGEX_HTTP, 'uri', PACKAGE_PREFIX)
 
@@ -1168,19 +1145,18 @@ module.exports = function (RED) {
 
   /**
    *  Play a specific track in queue. Queue must not be empty.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string/number} msg.payload position of track in queue. 1 ... queue length.
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
    * @param {boolean} [msg.sameVolume=true] shall all players play at same volume level.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupPlayTrack (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPlayTrack (msg, tsPlayer) {
     // Get the playerName
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
@@ -1215,13 +1191,12 @@ module.exports = function (RED) {
 
   /**
    *  Play tuneIn station. Optional set volume, use playerName.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload TuneIn id
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
    * @param {boolean} [msg.sameVolume=true] shall all players play at same volume level. 
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
@@ -1229,7 +1204,7 @@ module.exports = function (RED) {
    *          all from getGroupCurrent
    *          if msg.sameVolume === false and player == standalone because non sense.
    */
-  async function groupPlayTuneIn (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPlayTuneIn (msg, tsPlayer) {
     // Payload radio id is required
     const validatedRadioId = validRegex(msg, 'payload', REGEX_RADIO_ID, 'radio id',
       PACKAGE_PREFIX)
@@ -1258,16 +1233,15 @@ module.exports = function (RED) {
 
   /**
    *  Play previous track on given group of players.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupPreviousTrack (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupPreviousTrack (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname)
@@ -1278,17 +1252,16 @@ module.exports = function (RED) {
 
   /**
    *  Queue uri.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string/number}msg.payload valid uri
    * @param {string} [msg.playerName=using nodesonosPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupQueueUri (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupQueueUri (msg, tsPlayer) {
     // Payload uri is required.
     const validatedUri = validRegex(msg, 'payload', REGEX_ANYCHAR, 'uri', PACKAGE_PREFIX)
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
@@ -1301,11 +1274,10 @@ module.exports = function (RED) {
 
   /**
    *  Queue spotify uri on given group queue.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string/number} msg.payload valid uri from spotify
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * Valid examples
    * spotify:track:5AdoS3gS47x40nBNlNmPQ8
@@ -1320,7 +1292,7 @@ module.exports = function (RED) {
    * @throws any functions throws error and explicit throws
    */
   // eslint-disable-next-line max-len
-  async function groupQueueUriFromSpotify (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupQueueUriFromSpotify (msg, tsPlayer) {
     // Payload uri is required.
     const validatedUri = validRegex(msg, 'payload', REGEX_ANYCHAR, 'spotify uri', PACKAGE_PREFIX)
     if (!(validatedUri.startsWith('spotify:track:')
@@ -1340,18 +1312,17 @@ module.exports = function (RED) {
 
   /**
    *  Remove a number of tracks in queue.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string/number} msg.payload number of track in queue. 1 ... queue length.
    * @param {number/string} [msg.numberOfTracks=1] number of track 1 ... queue length.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupRemoveTracks (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupRemoveTracks (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
 
@@ -1383,18 +1354,17 @@ module.exports = function (RED) {
 
   /**
    *  Save SONOS queue to Sonos playlist.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload title of Sonos playlist.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
   // eslint-disable-next-line max-len
-  async function groupSaveQueueToSonosPlaylist (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupSaveQueueToSonosPlaylist (msg, tsPlayer) {
     // Payload title search string is required.
     const validatedTitle = validRegex(msg, 'payload', REGEX_ANYCHAR, 'title', PACKAGE_PREFIX)
 
@@ -1417,17 +1387,16 @@ module.exports = function (RED) {
 
   /**
    *  Group seek to specific time.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload hh:mm:ss time in song.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupSeek (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupSeek (msg, tsPlayer) {
     // Payload seek time is required.
     const validTime = validRegex(msg, 'payload', REGEX_TIME, 'seek time', PACKAGE_PREFIX)
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
@@ -1441,17 +1410,17 @@ module.exports = function (RED) {
 
   /**
    *  Group seek with delta time to specific time.
-   * @param {object} node not used
+
    * @param {object} msg incoming message
    * @param {string} msg.payload +/- hh:mm:ss time in song.
-   * @param {string} [msg.playerName=using nodesonosPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupSeekDelta (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupSeekDelta (msg, tsPlayer) {
     // Payload seek time is required.
     const validTime = validRegex(msg, 'payload', REGEX_TIME_DELTA, 'relative seek time',
       PACKAGE_PREFIX)
@@ -1466,17 +1435,16 @@ module.exports = function (RED) {
 
   /**
    *  Set group crossfade on|off.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload on|off.
-   * @param {string} [msg.playerName=using nodesonosPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupSetCrossfade (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupSetCrossfade (msg, tsPlayer) {
     // Payload crossfade sate is required.
     const newState = isOnOff(msg, 'payload', 'crosssfade state', PACKAGE_PREFIX)
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
@@ -1490,17 +1458,16 @@ module.exports = function (RED) {
 
   /**
    *  Set group mute state.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload on|off.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupSetMute (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupSetMute (msg, tsPlayer) {
     // Payload mute state is required.
     const newState = isOnOff(msg, 'payload', 'mute state', PACKAGE_PREFIX)
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
@@ -1514,17 +1481,16 @@ module.exports = function (RED) {
 
   /**
    *  Set group queuemode - queue must being activated and must not be empty.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload queue modes - may be mixed case
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupSetQueuemode (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupSetQueuemode (msg, tsPlayer) {
     // Payload queuemode is required.
     const newState = validRegex(msg, 'payload', REGEX_QUEUEMODES, 'queue mode', PACKAGE_PREFIX)
 
@@ -1561,17 +1527,16 @@ module.exports = function (RED) {
 
   /**
    *  Set group sleep timer.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload hh:mm:ss time in song.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupSetSleeptimer (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupSetSleeptimer (msg, tsPlayer) {
     // Payload sleep time is required.
     const validTime = validRegex(msg, 'payload', REGEX_TIME, 'timer duration',
       PACKAGE_PREFIX)
@@ -1587,17 +1552,16 @@ module.exports = function (RED) {
 
   /**
    *  Group set volume.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string/number} msg.payload new volume
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error
    */
-  async function groupSetVolume (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupSetVolume (msg, tsPlayer) {
     const newVolume = validToInteger(msg, 'payload', -100, +100, 'new volume', PACKAGE_PREFIX)
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
@@ -1610,16 +1574,15 @@ module.exports = function (RED) {
 
   /**
    *  Stop playing in that group, the specified player belongs to.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupStop (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupStop (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname)
@@ -1631,16 +1594,15 @@ module.exports = function (RED) {
 
   /**
    *  Toggle playback on given group of players.
-   * @param {object} node not used
    * @param {object} msg incoming message
-   * @param {string} [msg.playerName=using nodesonosPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function groupTogglePlayback (node, msg, nodesonosPlayer, tsPlayer) {
+  async function groupTogglePlayback (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const tsCoordinator = new SonosDevice(groupData.members[0].urlObject.hostname)
@@ -1652,10 +1614,9 @@ module.exports = function (RED) {
 
   /**
    *  Create a new group in household.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload csv list of playerNames, first will become coordinator
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} array of all group array of members :-)
    *
@@ -1666,7 +1627,7 @@ module.exports = function (RED) {
   // then just take that group and remove (first)/ add (second) the needed players.
   // else make the new coordinator stand alone and add all needed players.
   
-  async function householdCreateGroup (node, msg, nodesonosPlayer, tsPlayer) {
+  async function householdCreateGroup (msg, tsPlayer) {
     const validatedPlayerList = validRegex(msg, 'payload', REGEX_CSV,
       'player list', PACKAGE_PREFIX)
     const newGroupPlayerArray = validatedPlayerList.split(',')
@@ -1765,11 +1726,10 @@ module.exports = function (RED) {
   /**
    *  Create a stereo pair of players. Right one will be hidden! 
    * Is only supported for some type of SONOS player.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload - left player, will keep visible
    * @param {string} msg.playerNameRight - right player, will become invisible
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
@@ -1779,7 +1739,7 @@ module.exports = function (RED) {
    *
    */
   // eslint-disable-next-line max-len
-  async function householdCreateStereoPair (node, msg, nodesonosPlayer, tsPlayer) {
+  async function householdCreateStereoPair (msg, tsPlayer) {
     // Both player are required
     const playerLeft = validRegex(msg, 'payload', REGEX_ANYCHAR, 'player name left',
       PACKAGE_PREFIX)
@@ -1824,15 +1784,15 @@ module.exports = function (RED) {
 
   /**
    *  Get household groups. Ignore hidden player.
-   * @param {object} node not used
+
    * @param {object} msg incoming message
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise<Array<Array>>} array of all group array of members
    *
    * @throws any functions throws error and explicit throws
    */
-  async function householdGetGroups (node, msg, nodesonosPlayer, tsPlayer) {
+  async function householdGetGroups (msg, tsPlayer) {
     const payload = await xGetGroupsAll(tsPlayer)
 
     return { payload }
@@ -1840,7 +1800,6 @@ module.exports = function (RED) {
 
   /**
    *  Get SONOS playlists (limited 200, ObjectID SQ)
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {object} tsPlayer sonos-ts player
    *
@@ -1848,7 +1807,7 @@ module.exports = function (RED) {
    *
    * @throws any functions throws error and explicit throws
    */
-  async function householdGetSonosPlaylists (node, msg, nodesonosPlayer, tsPlayer) {
+  async function householdGetSonosPlaylists (msg, tsPlayer) {
     const payload = await xGetSonosPlaylists(tsPlayer, REQUESTED_COUNT_PLAYLISTS)
 
     return { payload }
@@ -1856,7 +1815,6 @@ module.exports = function (RED) {
 
   /**
    *  Remove Sonos playlist with given title. (impact on My Sonos and also Sonos playlist list)
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload title of Sonos playlist.
    * @param {boolean} [msg.ignoreNotExists] if missing assume true
@@ -1866,7 +1824,7 @@ module.exports = function (RED) {
    *
    * @throws any functions throws error and explicit throws
    */
-  async function householdRemoveSonosPlaylist (node, msg, nodesonosPlayer, tsPlayer) {
+  async function householdRemoveSonosPlaylist (msg, tsPlayer) {
     // Payload title search string is required.
     const validatedTitle = validRegex(msg, 'payload', REGEX_ANYCHAR, 'title',
       PACKAGE_PREFIX)
@@ -1901,16 +1859,15 @@ module.exports = function (RED) {
 
   /**
    *  Separate group in household.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function householdSeparateGroup (node, msg, nodesonosPlayer, tsPlayer) {
+  async function householdSeparateGroup (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     for (let i = 1; i < groupData.members.length; i++) { // Start with 1 - coordinator is last
@@ -1926,10 +1883,9 @@ module.exports = function (RED) {
 
   /**
    *  Separate a stereo pair of players. Right player will become visible again.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload - left SONOS-Playername, is visible
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
@@ -1937,7 +1893,7 @@ module.exports = function (RED) {
    *
    */
   // eslint-disable-next-line max-len
-  async function householdSeparateStereoPair (node, msg, nodesonosPlayer, tsPlayer) {
+  async function householdSeparateStereoPair (msg, tsPlayer) {
     // Player left is required
     const playerLeft = validRegex(msg, 'payload', REGEX_ANYCHAR, 'player name left',
       PACKAGE_PREFIX)
@@ -1992,10 +1948,9 @@ module.exports = function (RED) {
 
   /**
    *  Household test player connection
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload SONOS player name, required!!!!
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} true | false
    *
@@ -2005,7 +1960,7 @@ module.exports = function (RED) {
    * @throws any functions throws error and explicit throws
    */
   // eslint-disable-next-line max-len
-  async function householdTestPlayerOnline (node, msg, nodesonosPlayer, tsPlayer) {
+  async function householdTestPlayerOnline (msg, tsPlayer) {
     // Player name is required
     if (!xIsTruthyProperty(msg, ['payload'])) {
       throw new Error(`${PACKAGE_PREFIX} player name (msg.${msg.payload}) is missing/invalid`)
@@ -2035,14 +1990,13 @@ module.exports = function (RED) {
 
   /**
    *  Play notification on a joiner (in group) specified by sonosPlayer (default) or by playerName.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload notification uri.
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
    * @param {string} [msg.duration] duration of notification hh:mm:ss 
    * - default is calculation, if that fails then 00:00:05
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
@@ -2052,7 +2006,7 @@ module.exports = function (RED) {
    *  While playing a notification (start .. to end + 2 seconds)
    *     there should not be send another request to this player and the group shound be modified
    */
-  async function joinerPlayNotification (node, msg, nodesonosPlayer, tsPlayer) {
+  async function joinerPlayNotification (msg, tsPlayer) {
     // Payload notification uri is required.
     const validatedUri = validRegex(msg, 'payload', REGEX_ANYCHAR, 'uri', PACKAGE_PREFIX)
 
@@ -2097,17 +2051,16 @@ module.exports = function (RED) {
 
   /**
    *  Adjust player volume.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string/number} msg.payload -100 to +100 integer.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerAdjustVolume (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerAdjustVolume (msg, tsPlayer) {
     // Payload volume is required.
     const adjustVolume = validToInteger(msg, 'payload', -100, +100, 'adjust volume',
       PACKAGE_PREFIX)
@@ -2125,16 +2078,15 @@ module.exports = function (RED) {
 
   /**
    *  Player become coordinator of standalone group.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerBecomeStandalone (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerBecomeStandalone (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     // eslint-disable-next-line max-len
@@ -2149,16 +2101,15 @@ module.exports = function (RED) {
 
   /**
    *  Get player bass.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {payload: bas} type string -10 .. 10
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetBass (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetBass (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const payload = await executeActionV6(groupData.members[groupData.playerIndex].urlObject,
@@ -2170,10 +2121,9 @@ module.exports = function (RED) {
 
   /**
    *  Get player EQ data.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} object to update msg. msg.payload the Loudness state LED state on|off
    *
@@ -2181,7 +2131,7 @@ module.exports = function (RED) {
    *
    * EQ data are only available for specific players.
    */
-  async function playerGetEq (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetEq (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     // eslint-disable-next-line max-len
@@ -2209,7 +2159,7 @@ module.exports = function (RED) {
       // Can not happen
     }
 
-    let payload = await executeActionV6(nodesonosPlayer.urlObject,
+    let payload = await executeActionV6(tsPlayer.urlObject,
       '/MediaRenderer/RenderingControl/Control', 'GetEQ',
       args)
 
@@ -2225,16 +2175,15 @@ module.exports = function (RED) {
 
   /**
    *  Get player LED state.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise<string>} payload on or off
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetLed (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetLed (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     // returns On or Off
@@ -2246,16 +2195,15 @@ module.exports = function (RED) {
 
   /**
    *  Get player loudness.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} object to update msg. msg.payload the Loudness state LED state on|off
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetLoudness (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetLoudness (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const loudness = await executeActionV6(groupData.members[groupData.playerIndex].urlObject,
@@ -2267,16 +2215,15 @@ module.exports = function (RED) {
 
   /**
    *  Get mute state for given player.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {payload: muteState} on|off
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetMute (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetMute (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const payload = await xGetMutestate(groupData.members[groupData.playerIndex].urlObject)
@@ -2286,16 +2233,15 @@ module.exports = function (RED) {
 
   /**
    *  Get player properties.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} object to update msg. msg.payload the properties object
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetProperties (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetProperties (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     // eslint-disable-next-line max-len
@@ -2311,16 +2257,15 @@ module.exports = function (RED) {
 
   /**
    *  Get the SONOS-Queue of the specified player.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} object to update msg. msg.payload = array of queue items as object
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetQueue (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetQueue (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     // eslint-disable-next-line max-len
@@ -2333,16 +2278,15 @@ module.exports = function (RED) {
 
   /**
    *  Get the role and name of a player.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} object to update msg. msg.payload to role of player as string.
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetRole (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetRole (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     let role
@@ -2359,16 +2303,15 @@ module.exports = function (RED) {
 
   /**
    *  Get player treble.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise<string>} string -10 .. 10
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetTreble (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetTreble (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const payload = await executeActionV6(groupData.members[groupData.playerIndex].urlObject,
@@ -2380,16 +2323,15 @@ module.exports = function (RED) {
 
   /**
    *  Get volume of given player.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise<string>} range 0 .. 100
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerGetVolume (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerGetVolume (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const payload = await xGetVolume(groupData.members[groupData.playerIndex].urlObject)
@@ -2399,11 +2341,10 @@ module.exports = function (RED) {
 
   /**
    *  Join a group. The group is being identified in payload (or config node)
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload SONOS name of any player in the group
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
@@ -2413,7 +2354,7 @@ module.exports = function (RED) {
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerJoinGroup (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerJoinGroup (msg, tsPlayer) {
     // Payload a playername in group is required
     const validatedGroupPlayerName = validRegex(msg, 'payload', REGEX_ANYCHAR,
       'group player name', PACKAGE_PREFIX)
@@ -2429,7 +2370,7 @@ module.exports = function (RED) {
     if (
       groupDataJoiner.members[groupDataJoiner.playerIndex].playerName
       !== groupDataToJoin.members[0].playerName) {
-      // No check - always returns true. We use SetAVTransport as build in AddMember does not work
+      // No check - always returns true. We use SetAVTransport as AddMember does not work
       await executeActionV6(groupDataJoiner.members[groupDataJoiner.playerIndex].urlObject,
         '/MediaRenderer/AVTransport/Control', 'SetAVTransportURI',
         { 'InstanceID': 0, 'CurrentURI': coordinatorRincon, 'CurrentURIMetaData': '' })
@@ -2440,10 +2381,9 @@ module.exports = function (RED) {
 
   /**
    *  Leave a group - means become a standalone player.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
@@ -2453,7 +2393,7 @@ module.exports = function (RED) {
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerLeaveGroup (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerLeaveGroup (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     await executeActionV6(groupData.members[groupData.playerIndex].urlObject,
@@ -2465,19 +2405,18 @@ module.exports = function (RED) {
   
   /**
    *  Player play AVTransport uri: LineIn, TV
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload uri x-***:
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
-   * @param {string} [msg.playerName=using nodesonosPlayer] SONOS-Playername
-   * @param {object} nodesonosPlayer player with urlObject - as default
+   * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    *
    */
-  async function playerPlayAvtransport (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerPlayAvtransport (msg, tsPlayer) {
     // Payload uri is required: eg x-rincon-stream:RINCON_5CAAFD00223601400 for line in
     const validatedUri = validRegex(msg, 'payload', REGEX_ANYCHAR, 'uri', PACKAGE_PREFIX)
 
@@ -2500,18 +2439,17 @@ module.exports = function (RED) {
 
   /**
    *  Player play TV
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {number/string} [msg.volume] volume - if missing do not touch volume
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    *
    */
-  async function playerPlayTv (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerPlayTv (msg, tsPlayer) {
     // Validate msg.playerName, msg.volume
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
@@ -2552,17 +2490,16 @@ module.exports = function (RED) {
 
   /**
    *  Set bass.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string/number} msg.payload-10 to +10 integer.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerSetBass (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerSetBass (msg, tsPlayer) {
     // Payload volume is required.
     const newBass = validToInteger(msg, 'payload', -10, +10, 'set bass', PACKAGE_PREFIX)
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
@@ -2576,18 +2513,17 @@ module.exports = function (RED) {
 
   /**
    *  Set player EQ type
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.topic the lowercase, player.set.nightmode/subgain/dialoglevel
    * @param {string} msg.payload value on,off or -15 .. 15 in case of subgain
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerSetEQ (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerSetEQ (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     // eslint-disable-next-line max-len
@@ -2630,17 +2566,16 @@ module.exports = function (RED) {
 
   /**
    *  Set player led on|off.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload on|off
    * @param {string} [msg.playerName=using tslayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerSetLed (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerSetLed (msg, tsPlayer) {
     // Msg.state is required - convert to On Off
     const newState = (isOnOff(msg, 'payload', 'led state', PACKAGE_PREFIX) ? 'On' : 'Off')
 
@@ -2655,17 +2590,16 @@ module.exports = function (RED) {
 
   /**
    *  Set player loudness on|off.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload on|off
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerSetLoudness (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerSetLoudness (msg, tsPlayer) {
     // Msg.state is required
     const newState = isOnOff(msg, 'payload', 'loudness state', PACKAGE_PREFIX)
 
@@ -2680,17 +2614,16 @@ module.exports = function (RED) {
 
   /**
    *  Set mute for given player.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.payload on|off.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerSetMute (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerSetMute (msg, tsPlayer) {
     // Payload mute state is required.
     const newState = isOnOff(msg, 'payload', 'mute state', PACKAGE_PREFIX)
 
@@ -2703,17 +2636,16 @@ module.exports = function (RED) {
 
   /**
    *  Player set treble.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string/number} msg.payload -10 to +10 integer.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerSetTreble (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerSetTreble (msg, tsPlayer) {
     // Payload volume is required.
     const newTreble = validToInteger(msg, 'payload', -10, +10, 'set treble',
       PACKAGE_PREFIX)
@@ -2729,17 +2661,16 @@ module.exports = function (RED) {
 
   /**
    *  Set volume for given player.
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {number/string} msg.payload, integer 0 .. 100 integer.
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerSetVolume (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerSetVolume (msg, tsPlayer) {
     // Payload volume is required.
     const validatedVolume = validToInteger(msg, 'payload', 0, 100, 'volume', PACKAGE_PREFIX)
     const validatedPlayerName = validRegex(msg, 'playerName', REGEX_ANYCHAR,
@@ -2752,19 +2683,18 @@ module.exports = function (RED) {
 
   /**
    *  Test action V5
-   * @param {object} node not used
    * @param {object} msg incoming message
    * @param {string} msg.endpoint 
    * @param {string} msg.action
    * @param {object} msg.inArgs
    * @param {string} [msg.playerName=using tsPlayer] SONOS-Playername
-   * @param {object} tsPlayer sonos-ts player with urlObject - as default
+   * @param {object} tsPlayer sonos-ts player with JavaScript build-in URL urlObject - as default
    *
    * @returns {promise} {}
    *
    * @throws any functions throws error and explicit throws
    */
-  async function playerExecuteActionV6 (node, msg, nodesonosPlayer, tsPlayer) {
+  async function playerExecuteActionV6 (msg, tsPlayer) {
     const validated = await validatedGroupProperties(msg, PACKAGE_PREFIX)
     const groupData = await xGetGroupCurrent(tsPlayer, validated.playerName)
     const { endpoint, action, inArgs } = msg
